Update and reorganize React Native vendor code
