+++ packages/react-native-web/src/exports/FlatList/index.js
@@ -1,2 +1,11 @@
+/**
+ * Copyright (c) 2016-present, Nicolas Gallagher.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
 import UnimplementedView from '../../modules/UnimplementedView';
 export default UnimplementedView;
