+++ packages/react-native-web/src/exports/PanResponder/index.js
@@ -1,2 +1,2 @@
-import PanResponder from '../../vendor/PanResponder';
+import PanResponder from '../../vendor/react-native/PanResponder';
 export default PanResponder;
