+++ packages/react-native-web/src/exports/Touchable/Position.js
@@ -7,7 +7,7 @@
  * @noflow
  */
 
-import PooledClass from '../../vendor/PooledClass';
+import PooledClass from '../../vendor/react-native/PooledClass';
 
 const twoArgumentPooler = PooledClass.twoArgumentPooler;
