+++ packages/react-native-web/src/exports/VirtualizedList/index.js
@@ -1,2 +1,11 @@
-import VirtualizedList from '../../vendor/VirtualizedList';
+/**
+ * Copyright (c) 2016-present, Nicolas Gallagher.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+import VirtualizedList from '../../vendor/react-native/VirtualizedList';
 export default VirtualizedList;
